se cambiaron el nombre de carpetas y se creo userController.js
