+++ middlewares/deleteMiddleware.js
@@ -1,6 +1,6 @@
 const path = require('path');
 
-const usersFilePath = path.join(__dirname, '../src/data/users.json');
+const usersFilePath = path.join(__dirname, '../src/data/usersData.json');
 
 function deleteMiddlewire(req,res,next){
     res.redirect('/');
