+++ src/views/login.ejs
@@ -31,7 +31,7 @@
                 </div>
               </div>
               <div class="row">
-                <input type="submit" value="Submit" />
+                <input type="submit" value="Enviar" />
               </div>
             </div>
           </form>
