+++ src/views/registro.ejs
@@ -122,7 +122,7 @@
               </div>
             </div>
             <div class="row">
-              <input type="submit" value="Submit">
+              <input type="submit" value="Enviar">
             </div>
           </div>
         </form>
